backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pdfminer.six==20221105
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
pytesseract==0.3.10
Pillow==10.1.0
groq==0.4.2